requirements.txt
# DinoFlow Requirements

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
tqdm>=4.64.0

# HuggingFace Transformers (for SegFormer & DINOv3)
transformers>=4.45.0

# Optional: for visualization
# matplotlib>=3.5.0
# opencv-python>=4.5.0
wandb
